// src/declarations/file_manager/file_manager.did.d.ts
export type Asset = { File: FileResponse } | { Directory: DirectoryResponse };
export type AssetWithId = { File: bigint } | { Directory: bigint };
export interface DirectoryEntity {
  id: bigint;
  permission: Permission;
  updated_at: bigint;
  owner: [] | [Principal];
  name: string;
  created_at: bigint;
  parent_id: [] | [bigint];
  is_protected: boolean;
}
export interface DirectoryResponse {
  id: bigint;
  permission: Permission;
  updated_at: bigint;
  owner: [] | [Principal];
  name: string;
  created_at: bigint;
  children: Vec;
  parent_id: [] | [bigint];
  is_protected: boolean;
}
export interface FileResponse {
  id: bigint;
  permission: Permission;
  updated_at: bigint;
  sha256: Uint8Array | number[];
  owner: [] | [Principal];
  metadata: [] | [string];
  name: string;
  path: string;
  size: bigint;
  mime_type: string;
  created_at: bigint;
  parent_id: [] | [bigint];
  chunks: BigUint64Array | bigint[];
  is_protected: boolean;
  extension: string;
}
// ...
export type Permission = { Private: null } | { Origin: Array<string> } | { Public: null };
// ...
export type Vec = Array<
  | {
      File: {
        id: bigint;
        permission: { Private: null } | { Origin: Array<string> } | { Public: null };
        updated_at: bigint;
        sha256: Uint8Array | number[];
        owner: [] | [Principal];
        metadata: [] | [string];
        name: string;
        path: string;
        size: bigint;
        mime_type: string;
        created_at: bigint;
        parent_id: [] | [bigint];
        chunks: BigUint64Array | bigint[];
        is_protected: boolean;
        extension: string;
      };
    }
  | {
      Directory: {
        id: bigint;
        permission: { Private: null } | { Origin: Array<string> } | { Public: null };
        updated_at: bigint;
        owner: [] | [Principal];
        name: string;
        created_at: bigint;
        children: Vec;
        parent_id: [] | [bigint];
        is_protected: boolean;
      };
    }
>;